Conflicts:
	Sources/Internal/Android.mk
	Sources/Internal/Render/2D/Systems/RenderSystem2D.cpp
	Sources/Internal/UI/UIScreenTransition.cpp
	Sources/Internal/UI/UITextFieldAndroid.cpp

+++ Sources/Internal/Math/MathHelpers.h
@@ -204,7 +204,7 @@ namespace DAVA
 
     inline float32 Round(float32 value)
 	{
-		return (float32)((value > 0.0) ? floor(value+ 0.5) : ceil(value - 0.5));
+		return (value > 0.0f) ? floor(value + 0.5f) : ceil(value - 0.5f);
 	}
 
 } // end of namespace DAVA
